Use curl instead of wget.
I have ported this script from a wget dependency to a curl dependency.
The replacement I've made was changing all instances from "wget -c" to
"curl -C - -OL" which should have the same behavior but with a different
readout. Given that libwfut depends on libcurl and the curl utility is merely
a small frontend for libcurl, this made more sense to me. This should also
make it easier for OSX and minGW users to gain support in the future, as curl
is provided with OSX, but wget must be downloaded. curl is also slightly
easier to install for minGW.