Fix issue with CEGUI and Lua 5.2 on Fedora 20.
Lua 5.1 is provided as "lua-5.1.pc", and 5.2 as "lua.pc". This makes the
CEGUI pick up Lua 5.2 if this is installed, even though tolua++ is ABI
incompatible. Segfaults await.
We fix this by making sure that CEGUI also looks for "lua-5.1" before
looking for "lua.pc". Once this has been fixed in a future version of
CEGUI we should remove this.